+++ experiments/template-instantiation/MultiplyEnginePlainExplicit.cpp
@@ -7,6 +7,7 @@
 
 #include "MultiplyEnginePlainExplicit.h"
 #include <cassert>
+#include <cstdint>   // std::int32_t
 
 namespace Oblio {
 
@@ -15,9 +16,9 @@ Vector<Val> MultiplyEngine::Multiply(const Matrix<Val>& A,
                                      const Vector<Val>& x) const {
     assert(A.cols() == x.size());
     Vector<Val> y(A.rows());
-    for (std::size_t i = 0; i < A.rows(); ++i) {
+    for (std::int32_t i = 0; i < static_cast<std::int32_t>(A.rows()); ++i) {
         Val sum{0};
-        for (std::size_t j = 0; j < A.cols(); ++j) {
+        for (std::int32_t j = 0; j < static_cast<std::int32_t>(A.cols()); ++j) {
             sum += A(i, j) * x[j];
         }
         y[i] = sum;
